requirements.txt
Django==5.1.4
django-axes==6.1.1
PyPDF2==3.0.1
fpdf==1.7.2
django-allauth==65.3.1
google-auth==2.37.0
google-api-python-client==2.157.0
jwt==1.3.1
google_auth_oauthlib==1.2.1
python-dotenv==1.0.1